Refine time filtering for micro-precision on short ranges
- Added 30-minute and 2-hour options for more granular filtering
- Enhanced time display with hours + minutes (e.g. '2h 15m ago')
- Changed hours_old parameter to float to support fractional hours (0.5 for 30min)
- Added visual indicators for fresh articles:
  * 🔥 HOT badge for articles under 1 hour
  * ✨ NEW badge for articles under 2 hours
- Fresh articles get red pulsing border animation
- Reduced cache time to 10min for more frequent updates on short ranges
- More precise datetime calculations using total_seconds()

+++ src/aggregator.py
@@ -134,7 +134,7 @@ class NewsAggregator:
                           limit: int = 200,
                           category: str = None,
                           source: str = None,
-                          hours_old: int = None) -> List[Dict]:
+                          hours_old: float = None) -> List[Dict]:
         """Get recent articles with filtering"""
         return self.db.get_articles(
             limit=limit,
